+++ R/mfl_playerscores.R
@@ -11,15 +11,55 @@
 #'
 #' @examples
 #' \donttest{
-#' dlf_conn <- mfl_connect(2020, league_id = 37920)
-#' ff_playerscores(conn = dlf_conn, season = 2019, week = "YTD")
+#' sfb_conn <- mfl_connect(2020, league_id = 65443)
+#' ff_playerscores(conn = sfb_conn, season = 2019, week = "YTD")
 #' }
 #' @export
 
 ff_playerscores.mfl_conn <- function(conn, season, week, ...) {
-  player_scores <- tidyr::crossing(season = season, week = week) %>%
+  earliest_year <- ff_league(conn) %>%
+    dplyr::pull(years_active) %>%
+    stringr::str_extract("^[0-9]+")
+
+  raw_playerscores <- tidyr::crossing(season = season, week = week) %>%
     dplyr::transmute(playerscore = purrr::map2(.data$season, .data$week, .mfl_playerscore, conn = conn)) %>%
     tidyr::unnest(playerscore)
+
+  players <- dplyr::tibble(season = season) %>%
+    dplyr::mutate(
+      connection = purrr::map_if(
+        .x = .data$season,
+        .p = .data$season >= earliest_year,
+        .f = ~ mfl_connect(.x, conn$league_id),
+        .else = ~ mfl_connect(.x)
+      ),
+      players = purrr::map(
+        .data$connection,
+        ~ mfl_players(.x) %>%
+          dplyr::select("player_id", "player_name", "pos", "team")
+      ),
+      connection = NULL
+    ) %>%
+    tidyr::unnest("players")
+
+  player_scores <- raw_playerscores %>%
+    dplyr::left_join(
+      players,
+      by = c("season", "player_id")
+    ) %>%
+    dplyr::select(
+      dplyr::any_of(c(
+        "season",
+        "week",
+        "player_id",
+        "player_name",
+        "pos",
+        "team",
+        "points",
+        "is_available"
+      )),
+      dplyr::everything()
+    )
 
   return(player_scores)
 }
@@ -33,23 +73,19 @@ ff_playerscores.mfl_conn <- function(conn, season, week, ...) {
     purrr::pluck("content", "playerScores", "playerScore") %>%
     tibble::tibble() %>%
     tidyr::unnest_wider(1) %>%
-    dplyr::left_join(
-      dplyr::select(mfl_players(mfl_connect(season)), "player_id", "player_name", "pos", "team"),
-      by = c("id" = "player_id")
-    ) %>%
     dplyr::mutate(
       season = season,
       week = week,
       score = as.numeric(.data$score)
     ) %>%
-    dplyr::select("season",
-      "week",
-      "player_id" = "id",
-      "player_name",
-      "pos",
-      "team",
-      "points" = "score",
-      "is_available" = "isAvailable",
+    dplyr::select(
+      dplyr::any_of(c(
+        "season",
+        "week",
+        "player_id" = "id",
+        "points" = "score",
+        "is_available" = "isAvailable"
+      )),
       dplyr::everything()
     )