+++ R/sleeper_playerscores.R
@@ -3,12 +3,14 @@
 #' Sleeper PlayerScores
 #'
 #' Unfortunately, Sleeper has deprecated their player stats endpoint from their supported/open API.
-#' Adding some nflfastr-based ideas to the pipeline for future iterations.
+#' Please see `ff_scoringhistory()` for an alternative reconstruction.
 #'
 #' @param conn the list object created by \code{ff_connect()}
 #' @param ... other arguments (currently unused)
 #'
 #' @describeIn ff_playerscores Sleeper: Deprecated their open API endpoint for player scores
+#'
+#' @seealso ff_scoringhistory
 #'
 #' @export