+++ R/sleeper_scoring.R
@@ -14,11 +14,15 @@
 #'
 #' @export
 ff_scoring.sleeper_conn <- function(conn) {
-  scoring_rules <- glue::glue("league/{conn$league_id}") %>%
+  scoring_rules <-
+    glue::glue("league/{conn$league_id}") %>%
     sleeper_getendpoint() %>%
     purrr::pluck("content", "scoring_settings") %>%
     tibble::enframe(name = "event", value = "points") %>%
-    dplyr::mutate(points = as.numeric(.data$points) %>% round(3))
+    dplyr::mutate(points = as.numeric(.data$points) %>% round(3)) %>%
+    # Look in data-raw `DATASET` script to change the sleeper rule mappings
+    dplyr::inner_join(sleeper_rule_mapping, by = "event") %>%
+    dplyr::select("pos", "event", "points")
 
   return(scoring_rules)
 }